+++ app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
             android:name=".ui.onboarding.OnboardingActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="false" >
         </activity>
