+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FEF5DC"
-    tools:context=".MainActivity">
+    tools:context=".ui.main.MainActivity">
 
 
 <!--    <TextView-->
